added decimal to cost

+++ db/migrations/create_makersbnb.sql
@@ -1,2 +1,2 @@
 CREATE DATABASE "makersbnb";
-CREATE TABLE listings(id SERIAL PRIMARY KEY, name TEXT, cost int, address TEXT, description TEXT);
+CREATE TABLE listings(id SERIAL PRIMARY KEY, name TEXT, cost decimal (18 ,2), address TEXT, description TEXT);
