+++ db/migrations/create_makersbnb_test.sql
@@ -1,2 +1,2 @@
 CREATE DATABASE "makersbnb_test";
-CREATE TABLE listings(id SERIAL PRIMARY KEY, name TEXT, cost int, address TEXT, description TEXT);
+CREATE TABLE listings(id SERIAL PRIMARY KEY, name TEXT, cost decimal (18 ,2), address TEXT, description TEXT);
